+++ UniversityHousingSystem.Core/Features/CollegeDepartment/Queries/Results/GetAllCollegeDepartmentResponse.cs
@@ -4,7 +4,7 @@ namespace UniversityHousingSystem.Core.Features.Events.Queries.Results
 {
     public class GetAllCollegeDepartmentResponse
     {
-        public byte CollegeDepartmentId { get; set; }
+        public int CollegeDepartmentId { get; set; }
         public string Name { get; set; } = default!;
         public int CollegeId { get; set; }
